requirements.txt
alembic==1.10.4
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
databases==0.7.0
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
GeoAlchemy2==0.13.2
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
passlib==1.7.4
phonenumbers==8.13.11
psycopg2-binary==2.9.6
pydantic==1.10.7
PyJWT==2.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.48
SQLAlchemy-Utils==0.41.1
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
